WCM-29 fix return type

+++ shared/content-targeting-core/src/com/liferay/contenttargeting/service/impl/UserSegmentLocalServiceImpl.java
@@ -152,14 +152,14 @@ public class UserSegmentLocalServiceImpl
 	}
 
 	@Override
-	public long getUserSegmentsCount(long groupId)
+	public int getUserSegmentsCount(long groupId)
 		throws PortalException, SystemException {
 
 		return userSegmentPersistence.countByGroupId(groupId);
 	}
 
 	@Override
-	public long getUserSegmentsCount(long[] groupIds)
+	public int getUserSegmentsCount(long[] groupIds)
 		throws PortalException, SystemException {
 
 		return userSegmentPersistence.countByGroupId(groupIds);
